+++ app/javascript/react/pages/CalendarPage/CalendarPage.tsx
@@ -1,5 +1,5 @@
 import moment from "moment";
-import React, { useCallback, useEffect, useRef, useState } from "react";
+import React, { useEffect, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
 
 import { Graph } from "../../components/Graph";
@@ -12,25 +12,28 @@ import { FixedStreamStationHeader } from "../../components/molecules/FixedStream
 import { ThresholdsConfigurator } from "../../components/ThresholdConfigurator";
 import { ResetButton } from "../../components/ThresholdConfigurator/ThresholdButtons/ResetButton";
 import { ThresholdButtonVariant } from "../../components/ThresholdConfigurator/ThresholdButtons/ThresholdButton";
-import { selectFixedStreamShortInfo } from "../../store/fixedStreamSelectors";
+import { UniformDistributionButton } from "../../components/ThresholdConfigurator/ThresholdButtons/UniformDistributionButton";
+
 import {
   fetchFixedStreamById,
   selectFixedData,
 } from "../../store/fixedStreamSlice";
 import { useAppDispatch, useAppSelector } from "../../store/hooks";
+import { selectFixedStreamShortInfo } from "../../store/fixedStreamSelectors";
 import {
   fetchNewMovingStream,
   movingData,
 } from "../../store/movingCalendarStreamSlice";
 import { setDefaultThresholdsValues } from "../../store/thresholdSlice";
+
 import { SessionTypes } from "../../types/filters";
-import { UrlParamsTypes, useMapParams } from "../../utils/mapParamsHandler";
+
+import { useMapParams } from "../../utils/mapParamsHandler";
 import { formatTime } from "../../utils/measurementsCalc";
 import useMobileDetection from "../../utils/useScreenSizeDetection";
+import { useCalendarBackNavigation } from "../../hooks/useBackNavigation";
+
 import * as S from "./CalendarPage.style";
-import { UniformDistributionButton } from "../../components/ThresholdConfigurator/ThresholdButtons/UniformDistributionButton";
-import { urls } from "../../const/urls";
-import { useNavigate } from "react-router-dom";
 
 interface CalendarPageProps {
   children: React.ReactNode;
@@ -40,15 +43,10 @@ const CalendarPage: React.FC<CalendarPageProps> = ({ children }) => {
   const dispatch = useAppDispatch();
   const isMobile = useMobileDetection();
   const { t } = useTranslation();
-  const navigate = useNavigate();
 
-  const {
-    unitSymbol,
-    revertUserSettingsAndResetIds,
-    searchParams,
-    currentUserSettings,
-    previousUserSettings,
-  } = useMapParams();
+  const handleCalendarGoBack = useCalendarBackNavigation();
+
+  const { unitSymbol } = useMapParams();
 
   const { streamId } = useMapParams();
 
@@ -76,29 +74,13 @@ const CalendarPage: React.FC<CalendarPageProps> = ({ children }) => {
     streamId && dispatch(fetchFixedStreamById(streamId));
   }, []);
 
-  const handleBackNavigation = useCallback(() => {
-    const newSearchParams = new URLSearchParams(searchParams.toString());
-    newSearchParams.set(
-      UrlParamsTypes.previousUserSettings,
-      currentUserSettings
-    );
-    newSearchParams.set(
-      UrlParamsTypes.currentUserSettings,
-      previousUserSettings
-    );
-
-    isMobile && newSearchParams.delete(UrlParamsTypes.streamId);
-    isMobile && newSearchParams.delete(UrlParamsTypes.sessionId);
-    navigate(`${urls.reactMap}?${newSearchParams.toString()}`);
-  }, [currentUserSettings, navigate, previousUserSettings, searchParams]);
-
   useEffect(() => {
-    window.addEventListener("popstate", handleBackNavigation);
+    window.addEventListener("popstate", handleCalendarGoBack);
 
     return () => {
-      window.removeEventListener("popstate", handleBackNavigation);
+      window.removeEventListener("popstate", handleCalendarGoBack);
     };
-  }, [handleBackNavigation]);
+  }, [handleCalendarGoBack]);
 
   useEffect(() => {
     const formattedEndMoment = moment(streamEndTime, "YYYY-MM-DD");
